+++ RevenueCat/Scripts/PurchaserInfo.cs
@@ -3,9 +3,10 @@ using System.Collections.Generic;
 using System.Diagnostics.CodeAnalysis;
 using JetBrains.Annotations;
 using RevenueCat.SimpleJSON;
+using static RevenueCat.Utilities;
 
 public partial class Purchases
-{    
+{
     /*
      * PurchaserInfo encapsulate the current status of subscriber. 
      * Use it to determine which entitlements to unlock, typically by checking 
@@ -29,7 +30,7 @@ public partial class Purchases
         public Dictionary<string, DateTime> AllPurchaseDates;
         [CanBeNull] public string OriginalApplicationVersion;
         [CanBeNull] public string ManagementURL;
-
+        public List<Transaction> NonSubscriptionTransactions;
 
         public PurchaserInfo(JSONNode response)
         {
@@ -39,17 +40,20 @@ public partial class Purchases
             {
                 ActiveSubscriptions.Add(subscription);
             }
+
             AllPurchasedProductIdentifiers = new List<string>();
             foreach (JSONNode productIdentifier in response["allPurchasedProductIdentifiers"])
             {
                 AllPurchasedProductIdentifiers.Add(productIdentifier);
             }
-            
+
             FirstSeen = FromUnixTime(response["firstSeenMillis"].AsLong);
             OriginalAppUserId = response["originalAppUserId"];
             RequestDate = FromUnixTime(response["requestDateMillis"].AsLong);
-            OriginalPurchaseDate = FromOptionalUnixTime(response["originalPurchaseDateMillis"].AsLong);
-            LatestExpirationDate = FromOptionalUnixTime(response["latestExpirationDateMillis"].AsLong);
+            OriginalPurchaseDate =
+                FromOptionalUnixTime(response["originalPurchaseDateMillis"].AsLong);
+            LatestExpirationDate =
+                FromOptionalUnixTime(response["latestExpirationDateMillis"].AsLong);
             ManagementURL = response["managementURL"];
             AllExpirationDates = new Dictionary<string, DateTime?>();
             foreach (var keyValue in response["allExpirationDatesMillis"])
@@ -70,26 +74,16 @@ public partial class Purchases
             foreach (var keyValue in response["allPurchaseDatesMillis"])
             {
                 AllPurchaseDates.Add(keyValue.Key, FromUnixTime(keyValue.Value.AsLong));
-            } 
+            }
 
             OriginalApplicationVersion = response["originalApplicationVersion"];
-        } 
-        
-        private static DateTime? FromOptionalUnixTime(long unixTime)
-         {
-            DateTime? value = null;
-             if (unixTime != 0L) { 
-                value = FromUnixTime(unixTime);
-             }
-             return value;
-         }
-
-        private static DateTime FromUnixTime(long unixTime)
-        {
-            return Epoch.AddSeconds(unixTime);
+            
+            NonSubscriptionTransactions = new List<Transaction>();
+            foreach (JSONNode transactionResponse in response["nonSubscriptionTransactions"])
+            {
+                NonSubscriptionTransactions.Add(new Transaction(transactionResponse));
+            }
         }
-
-        private static readonly DateTime Epoch = new DateTime(1970, 1, 1, 0, 0, 0, DateTimeKind.Utc);
 
         public override string ToString()
         {
@@ -104,7 +98,9 @@ public partial class Purchases
                    $"{nameof(AllPurchaseDates)}: {AllPurchaseDates}, " +
                    $"{nameof(OriginalPurchaseDate)}: {OriginalPurchaseDate}, " +
                    $"{nameof(ManagementURL)}: {ManagementURL}, " +
+                   $"{nameof(NonSubscriptionTransactions)}: {NonSubscriptionTransactions}, " +
                    $"{nameof(OriginalApplicationVersion)}: {OriginalApplicationVersion}";
         }
+
     }
 }