+++ scripts/create-unity-package.sh
@@ -1,7 +1,6 @@
 #!/usr/bin/env bash
 PROJECT="$PWD/Subtester"
 PACKAGE="$PWD/Purchases.unitypackage"
-PACKAGE_OBSERVER_MODE="$PWD/Purchases_NoInAppBillingService.unitypackage"
 FOLDERS_TO_EXPORT=$(cd $PROJECT; find Assets/RevenueCat/* Assets/PlayServicesResolver Assets/ExternalDependencyManager -type d -prune ! -name ObserverMode.xml)
 PLUGINS_FOLDER="$PWD/RevenueCat/Plugins"
 
@@ -60,34 +59,7 @@ if [[ $? -ne 0 ]]; then
   echo "😱 Unity shouldn't be running when executing the script."
   exit $return_code
 else
-  echo "🚀 Exporting finished for the regular package! Path: $PACKAGE" 
+  echo "🚀 Exporting finished! Path: $PACKAGE" 
 fi
-
-echo "📦 Moving com.android.billingclient.billing-no-service.aar to Revenuecat/Plugins/Android"
-cp $PWD/ObserverMode/com.android.billingclient.billing-no-service.aar $PLUGINS_FOLDER/Android
-
-echo "📦 Modifying RevenueCatDependencies.xml to use observer mode dependencies"
-mkdir $PWD/Temp
-mv $PLUGINS_FOLDER/Editor/RevenueCatDependencies.xml $PWD/Temp/RevenueCatDependencies.xml.bck
-cp $PWD/ObserverMode/RevenueCatDependencies.xml $PLUGINS_FOLDER/Editor/RevenueCatDependencies.xml
-
-echo "📦 Creating Purchases_NoInAppBillingService.unitypackage, this may take a minute."
-$UNITY_BIN -gvh_disable -projectPath $PROJECT -force-free -quit -batchmode -logFile exportlog.txt \
-    -importPackage $PROJECT/external-dependency-manager-*.unitypackage \
-    -exportPackage $FOLDERS_TO_EXPORT $PACKAGE_OBSERVER_MODE >& /dev/null
-
-if [[ $? -ne 0 ]]; then
-  echo "😱 Exporting has failed! Checkout exportlogs.txt"
-  echo "😱 Unity shouldn't be running when executing the script."
-  exit $return_code
-else
-  echo "🚀 Exporting finished for the observer mode package! Path: $PACKAGE_OBSERVER_MODE" 
-fi
-
-echo "📦 Cleaning up..."
-rm $PLUGINS_FOLDER/Android/com.android.billingclient.billing-no-service.aar
-rm $PLUGINS_FOLDER/Editor/RevenueCatDependencies.xml
-mv $PWD/Temp/RevenueCatDependencies.xml.bck $PLUGINS_FOLDER/Editor/RevenueCatDependencies.xml
-rm -rf $PWD/Temp
 
 open .