+++ CHANGELOG.md
@@ -1,3 +1,14 @@
+## 2.3.0
+
+- Bumped common files to 1.4.4 [Changelog here](https://github.com/RevenueCat/purchases-hybrid-common/releases/tag/1.4.4)
+- Bumped iOS to 3.7.2 [Changelog here](https://github.com/RevenueCat/purchases-ios/releases/tag/3.7.2) 
+- Bumped Android to 3.5.2 [Changelog here](https://github.com/RevenueCat/purchases-android/releases/tag/3.5.2)
+- Added a new property `NonSubscriptionTransactions` in `PurchaserInfo` to better manage non-subscriptions
+- Attribution V2:
+ - Deprecated addAttribution in favor of setAdjustId, setAppsflyerId, setFbAnonymousId, setMparticleId.
+ - Added support for OneSignal via setOnesignalId
+ - Added setMediaSource, setCampaign, setAdGroup, setAd, setKeyword, setCreative, and collectDeviceIdentifiers
+
 ## 2.2.1
 
 - Fixes duplicated files error in iOS.